+++ tests/feature/test_gherkin_terminal_reporter.py
@@ -225,7 +225,7 @@ def test_step_parameters_should_be_replaced_by_their_values(testdir):
         )
     )
 
-    result = testdir.runpytest("--gherkin-terminal-reporter", "--gherkin-terminal-reporter-expanded", "-vv")
+    result = testdir.runpytest("--gherkin-terminal-reporter", "-vv")
     result.assert_outcomes(passed=1, failed=0)
     result.stdout.fnmatch_lines("*Scenario: Scenario example 2")
     result.stdout.fnmatch_lines("*Given there are {start} cucumbers".format(**example))